refactor: HTTP 헤더 enum 명 변경
