+++ tomcat/src/main/java/org/apache/coyote/request/HttpRequest.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static org.apache.coyote.common.HttpHeader.CONTENT_LENGTH;
+import static org.apache.coyote.common.HeaderType.CONTENT_LENGTH;
 
 public class HttpRequest {
